requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-filter==24.3
django-jalali==7.2.0
django-rest-swagger==2.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
firebase-admin==6.6.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
ippanel==2.0.7
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.7
MarkupSafe==2.1.5
msgpack==1.1.0
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rsa==4.9
service-identity==24.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
